requirements.txt
Flask
protobuf
python-dotenv
google-generativeai
python-dotenv
requests